authorization completed

+++ routers/todos.py
@@ -99,5 +99,5 @@ async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = P
     if todo_model is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Todo not found")
     
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).delete()
     db.commit()
